macOS/Clang: fix Clang errors uncovered by upstream rebase
Builds the full CMake target set (backend + frontend + wfrontend +
designer + enterprise + launcher + daemon + codeRunner + classChecker +
wenterprise-server) cleanly on macOS arm64 / AppleClang 21 after rebase
onto upstream v1.3.0.

Backend / compiler — Clang strictness, MSVC accepted them implicitly:

- procUnitLinq.cpp: m_idx and count typed as 'long'. On macOS arm64
  'long' is 64-bit, which Clang cannot disambiguate between ibNumber's
  int64_t and uint64_t constructors. Change to int64_t — MSVC x64 long
  is 32-bit so it goes cleanly to ibNumber(int), but every other
  platform needs an explicit 64-bit type.
- valueArray.cpp: same root cause on a (long) cast — change to int64_t.
- commonObjectRefQuery.cpp::GenerateNextIdentifier: assigning a 'long'
  from wxString::ToLong(&parsed) into ibNumber maxFound is ambiguous on
  Clang; cast to int64_t at the assignment site.
- translateCode.cpp::ibLexem accessors: 'cond ? wxString : wxEmptyString'
  is ambiguous (wxString ↔ wxChar* both convertible). Switch to
  if/else with a function-static empty wxString to give both branches
  the same type.

Editor / IntelliSense API — non-const reference to a temporary:

- codeEditorInterpreter.h: FindVariable/FindFunction had default
  parameter 'ibValue& valContext = ibValue()'. Clang rejects binding a
  non-const reference to a default-constructed temporary. Change to
  pointer 'ibValue* valContext = nullptr'; impl updated to null-check
  before write; 2 callsites updated to pass &valContext.

Tests:

- test_compiler.cpp: was reaching cc.m_listLexem directly (protected).
  Add a GetLexemList() const accessor on ibTranslateCode and route
  the tests through it.

Build configuration:

- src/engine/codeRunner/CMakeLists.txt: link 'frontend' lib. Upstream
  lifted the code editor into frontend.dll (ac2fa26d), so codeRunner
  no longer satisfies ibCodeEditor::{ShowMethods,ShowGotoLine,
  DecreaseIndent,IncreaseIndent,...} from backend alone.
- src/engine/frontend/CMakeLists.txt: WFRONTEND_SOURCES — extend with
  tableBox.cpp / tableBox_res.cpp / tableBoxColumn.cpp /
  tableBoxColumn_res.cpp. formObject.cpp now references
  ibValueModelTableBox{,Column}::ms_classInfo, RefreshModel, GetIcon,
  GetIconGroup; these live in those four files. tableBoxColumnRenderer
  stays out because its header includes dataview.h whose enums
  conflict with the web-stub enums in tableBox.h's #ifndef OES_USE_WEB
  branch.

+++ src/engine/backend/metaCollection/partial/commonObjectRefQuery.cpp
@@ -340,7 +340,7 @@ ibValue ibValueRecordDataObjectRef::GenerateNextIdentifier(ibValueMetaObjectAttr
 						if (maxCode.length() > strPrefix.length()) {
 							long parsed = 0;
 							maxCode.Mid(strPrefix.length()).ToLong(&parsed);
-							if (parsed > 0) maxFound = parsed;
+							if (parsed > 0) maxFound = static_cast<int64_t>(parsed);
 						}
 					}
 					db->CloseResultSet(rs);
